Task 1/requirements/requirements.txt
pillow==11.1.0
numpy==2.0.2
scikit-learn==1.6.1
tensorflow==2.18.0
matplotlib==3.10.0